backend/requirements.txt
#Python version: 3.12.10
accelerate==1.12.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.0
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.3
black==25.12.0
build==1.3.0
cachetools==6.2.4
certifi==2026.1.4
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.4
chromadb==1.4.0
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.3
coverage==7.13.1
cryptography==46.0.3
cycler==0.12.1
dataclasses-json==0.6.7
Deprecated==1.3.1
distlib==0.4.0
distro==1.9.0
dnspython==2.8.0
durationpy==0.10
ecdsa==0.19.1
effdet==0.4.1
email-validator==2.3.0
emoji==2.15.0
fastapi==0.128.0
filelock==3.20.2
filetype==1.2.0
flatbuffers==25.12.19
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.12.0
google-api-core==2.28.1
google-auth==2.45.0
google-cloud-vision==3.11.0
googleapis-common-protos==1.72.0
greenlet==3.3.0
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
humanfriendly==10.0
identify==2.6.15
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
iniconfig==2.3.0
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
kubernetes==34.1.0
langchain==1.2.0
langchain-chroma==1.1.0
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.6
langchain-mongodb==0.9.0
langchain-openai==1.1.6
langchain-tavily==0.2.16
langchain-text-splitters==1.1.0
langchain-unstructured==1.0.1
langdetect==1.0.9
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-checkpoint-mongodb==0.3.0
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.1
langsmith==0.6.0
lark==1.3.1
librt==0.7.7
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.4
mmh3==5.2.0
motor==3.7.1
mpmath==1.3.0
multidict==6.7.0
mypy==1.19.1
mypy_extensions==1.1.0
networkx==3.6.1
nltk==3.9.2
nodeenv==1.10.0
numpy==2.2.6
oauthlib==3.3.1
olefile==0.47
omegaconf==2.3.0
onnx==1.20.0
onnxruntime==1.23.2
openai==2.14.0
opencv-python==4.12.0.88
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.5
ormsgpack==1.12.1
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
pdf2image==1.17.0
pdfminer.six==20251230
pi_heif==1.1.1
pikepdf==10.1.0
pillow==12.1.0
platformdirs==4.5.1
pluggy==1.6.0
posthog==5.4.0
pre_commit==4.5.1
propcache==0.4.1
proto-plus==1.27.0
protobuf==6.33.2
psutil==7.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.3
pycocotools==2.0.11
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
pymongo==4.15.5
pymongo-search-utils==0.1.0
PyMuPDF==1.26.7
pyparsing==3.3.1
pypdf==6.5.0
pypdfium2==5.3.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-iso639==2025.11.16
python-jose==3.5.0
python-magic==0.4.27
python-magic-bin==0.4.14
python-multipart==0.0.21
python-oxmsg==0.0.2
pytokens==0.3.0
pytz==2025.2
PyYAML==6.0.3
RapidFuzz==3.14.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.30.0
rsa==4.9.1
ruff==0.14.10
safetensors==0.7.0
scipy==1.16.3
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.1
SQLAlchemy==2.0.45
starlette==0.50.0
sympy==1.14.0
tavily-python==0.7.17
tenacity==9.1.2
tiktoken==0.12.0
timm==1.0.22
tokenizers==0.22.2
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
transformers==4.57.3
typer==0.21.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
unstructured==0.18.24
unstructured-client==0.42.6
unstructured.pytesseract==0.3.15
unstructured_inference==1.1.4
urllib3==2.3.0
uuid_utils==0.12.0
uvicorn==0.40.0
virtualenv==20.35.4
watchfiles==1.1.1
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
wrapt==2.0.1
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0